Crate libobs_window_helper

OBS Window Helper

This crate provides necessary information about windows that could be used so they can be captured with the window_capture or game_capture source in OBS.
The function you probably want to use is get_all_windows which returns a list of WindowInfo structs.

Structs

ProcessInfo

WindowInfo

Represents information about a window.

Enums

WindowSearchMode

Functions

get_all_windows

Retrieves information about all windows based on the specified search mode and game check flag.

get_thread_proc_id

get_window_info

Retrieves the OBS window information associated with the given window handle.

is_blacklisted_window

is_microsoft_internal_exe

is_window_valid